refactor(users -> auth): move auth logic to separate module
