+++ src/auth/index.ts
@@ -1,0 +1,3 @@
+import router from './auth-router.js'
+
+export {router as authRoutes};
