+++ src/main.ts
@@ -4,6 +4,7 @@ import {patientsRoutes} from "./patients/index.js";
 import {errorHandler} from "./shared/middleware/error-handler.js";
 import {appointmentsRoutes} from "./appointments/index.js";
 import {usersRoutes} from "./users/index.js";
+import {authRoutes} from "./auth/index.js";
 
 const app: Application = express()
 const PORT = process.env.PORT
@@ -14,6 +15,7 @@ app.use('/doctors', doctorsRoutes)
 app.use('/patients', patientsRoutes)
 app.use('/appointments', appointmentsRoutes)
 app.use('/users', usersRoutes)
+app.use('/auth', authRoutes)
 
 app.use(errorHandler)
