+++ src/users/users-controller.ts
@@ -1,15 +1,8 @@
 import {Request, Response} from "express";
 import {
     UsersParams,
-    CreateUserBody,
     UpdateUserBody,
-    UserPublic,
-    UserCredentials,
-    AuthTokens,
-    ChangePasswordBody,
-    ResetPasswordBody,
-    ResetPasswordQuery,
-    RequestResetPasswordBody
+    UserPublic
 } from "./types.js";
 import {Service} from "./service.js";
 
@@ -23,41 +16,6 @@ export const getAll = async (req: Request, res: Response<UserPublic[]>): Promise
 export const getById = async (req: Request<UsersParams>, res: Response<UserPublic | undefined>): Promise<void> => {
     const user = await Service.getById(req)
     res.status(200).send(user)
-}
-
-export const register = async (req: Request<{}, unknown, CreateUserBody>, res: Response<UserPublic>): Promise<void> => {
-    const newUser = await Service.register(req);
-    res.cookie('refreshToken', newUser.tokens.refreshToken, {
-        secure: true,
-        httpOnly: true,
-        maxAge: 30 * 24 * 60 * 60 * 1000
-    })
-    res.status(201).send(newUser.userData)
-}
-
-export const login = async (req: Request<{}, unknown, UserCredentials>, res: Response<AuthTokens>): Promise<void> => {
-    const tokens = await Service.login(req)
-    res.status(201).send(tokens)
-}
-
-export const logout = async(req: Request, res: Response): Promise<void> => {
-    await Service.logout(req);
-    res.status(200)
-}
-
-export const changePassword = async (req: Request<{}, unknown, ChangePasswordBody>, res: Response<{ message: string }>): Promise<void> => {
-    await Service.changePassword(req)
-    res.status(200).send({message: "Password was changed successfully"})
-}
-
-export const requestResetPassword = async (req: Request<{}, unknown, RequestResetPasswordBody>, res: Response<{message: string}>): Promise<void> => {
-    await Service.requestResetPassword(req)
-    res.status(200).send({message: "Link for changing password was sent to email"})
-}
-
-export const resetPassword = async (req: Request<{}, unknown, ResetPasswordBody, ResetPasswordQuery>, res: Response<{message: string}>): Promise<void> => {
-    await Service.resetPassword(req)
-    res.status(200).send({message: "Password was changed successfully"})
 }
 
 export const update = async (req: Request<UsersParams, unknown, UpdateUserBody>, res: Response<UserPublic>): Promise<void> => {
